pyproject.toml
[tool.poetry]
name = "vk_bot"
version = "0.1.0"
description = ""
authors = ["Shamil Khastiev <s.hastiev@innopolis.university>"]

[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^2.20.0"
vk-api = "^11.9.9"
opencv-python = "^4.6.0"
Pillow = "^9.2.0"
fire = "^0.4.0"
face-recognition = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
